requirements.txt
streamlit>=1.33
pympi-ling
praatio